pyproject.toml
[tool.poetry]
name = "siege-spider-client"
version = "0.1.0"
description = ""
authors = ["David Peterson <34923485+Haste171@users.noreply.github.com>"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.115.12"
uvicorn = "^0.34.2"
requests = "^2.32.3"
websocket-client = "^1.8.0"

[tool.poetry.group.dev.dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
